refactor: reorganize and implement getText function with SelectorStrategy
